+++ src/users/users.service.ts
@@ -1,9 +1,9 @@
-import { BadRequestException, Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { User, UserSchema } from './schema/user.schema';
 import { InjectModel } from '@nestjs/mongoose';
-import mongoose, { Model } from 'mongoose';
+import mongoose, { isValidObjectId, Model } from 'mongoose';
 @Injectable()
 export class UsersService {
   constructor(@InjectModel(User.name) private userModel: Model<User>) {}
@@ -20,15 +20,24 @@ export class UsersService {
     return this.userModel.find();
   }
 
-  findOne(id: mongoose.Schema.Types.ObjectId) {
-    return this.userModel.findById(id);
+  async findOne(id: mongoose.Schema.Types.ObjectId) {
+    if (!isValidObjectId(id)) throw new BadRequestException('Invalid mongo ID');
+    const user = await this.userModel.findById(id);
+    if (!user) throw new NotFoundException('User not found');
+    return user;
   }
 
-  update(id: mongoose.Schema.Types.ObjectId, updateUserDto: UpdateUserDto) {
-    return this.userModel.findByIdAndUpdate(id, updateUserDto);
+  async update(id: mongoose.Schema.Types.ObjectId,updateUserDto: UpdateUserDto) {
+    if (!isValidObjectId(id)) throw new BadRequestException('Invalid mongo ID');
+    const updatedUser = await this.userModel.findByIdAndUpdate(id,updateUserDto,{ new: true });
+    if (!updatedUser) throw new NotFoundException('User not found');
+    return updatedUser;
   }
 
-  remove(id: mongoose.Schema.Types.ObjectId) {
-    return this.userModel.findByIdAndDelete(id);
+  async remove(id: mongoose.Schema.Types.ObjectId) {
+    if (!isValidObjectId(id)) throw new BadRequestException('Invalid mongo ID');
+    const deletedUser = await this.userModel.findByIdAndDelete(id);
+    if (!deletedUser) throw new NotFoundException('User not found');
+    return deletedUser;
   }
 }
